+++ contracts/orchestrators/LiquidityOrchestrator.sol
@@ -47,7 +47,7 @@ contract LiquidityOrchestrator is Ownable, ILiquidityOrchestrator {
     /* -------------------------------------------------------------------------- */
 
     /// @notice Last processed epoch counter from Internal States Orchestrator
-    uint256 public lastProcessedEpoch;
+    uint16 public lastProcessedEpoch;
 
     /// @notice Upkeep phase
     LiquidityUpkeepPhase public currentPhase;
@@ -139,7 +139,7 @@ contract LiquidityOrchestrator is Ownable, ILiquidityOrchestrator {
 
     // TODO: docs when implemented.
     function checkUpkeep(bytes calldata) external view override returns (bool upkeepNeeded, bytes memory performData) {
-        uint256 currentEpoch = internalStatesOrchestrator.epochCounter();
+        uint16 currentEpoch = internalStatesOrchestrator.epochCounter();
         if (currentEpoch > lastProcessedEpoch && config.isSystemIdle()) {
             upkeepNeeded = true;
             // TODO: same as internal states orchestrator, use bytes4 and encodePacked.
@@ -155,7 +155,7 @@ contract LiquidityOrchestrator is Ownable, ILiquidityOrchestrator {
     // TODO: refacto for scalability, same as internal states orchestrator.
     // TODO: docs when implemented.
     function performUpkeep(bytes calldata) external override onlyAutomationRegistry {
-        uint256 currentEpoch = internalStatesOrchestrator.epochCounter();
+        uint16 currentEpoch = internalStatesOrchestrator.epochCounter();
         if (currentEpoch <= lastProcessedEpoch) {
             return;
         }
@@ -176,7 +176,7 @@ contract LiquidityOrchestrator is Ownable, ILiquidityOrchestrator {
             .getSellingOrders();
 
         // Sell before buy, avoid undercollateralization risk.
-        for (uint256 i = 0; i < sellingTokens.length; i++) {
+        for (uint16 i = 0; i < sellingTokens.length; i++) {
             address token = sellingTokens[i];
             uint256 amount = sellingAmounts[i];
             _executeSell(token, amount);
@@ -191,15 +191,15 @@ contract LiquidityOrchestrator is Ownable, ILiquidityOrchestrator {
 
         (address[] memory buyingTokens, uint256[] memory buyingAmounts) = internalStatesOrchestrator.getBuyingOrders();
 
-        for (uint256 i = 0; i < buyingTokens.length; i++) {
+        for (uint16 i = 0; i < buyingTokens.length; i++) {
             address token = buyingTokens[i];
             uint256 amount = buyingAmounts[i];
             _executeBuy(token, amount);
         }
 
         address[] memory transparentVaults = config.getAllOrionVaults(EventsLib.VaultType.Transparent);
-        uint256 length = transparentVaults.length;
-        for (uint256 i = 0; i < length; i++) {
+        uint16 length = uint16(transparentVaults.length);
+        for (uint16 i = 0; i < length; i++) {
             IOrionTransparentVault vault = IOrionTransparentVault(transparentVaults[i]);
             // TODO: implement.
             // vault.updateVaultState(?, ?);
@@ -208,8 +208,8 @@ contract LiquidityOrchestrator is Ownable, ILiquidityOrchestrator {
         // TODO: to updateVaultState of encrypted vaults, get the encrypted sharesPerAsset executed by the liquidity
 
         address[] memory encryptedVaults = config.getAllOrionVaults(EventsLib.VaultType.Encrypted);
-        length = encryptedVaults.length;
-        for (uint256 i = 0; i < length; i++) {
+        length = uint16(encryptedVaults.length);
+        for (uint16 i = 0; i < length; i++) {
             IOrionEncryptedVault vault = IOrionEncryptedVault(encryptedVaults[i]);
             // TODO: implement.
             // vault.updateVaultState(?, ?);
